fix(tests): resolve SqlValue::Real type and reqwest version issues
- Update conversion_function_tests.rs to use f64 for real_lit() after
  SqlValue::Real changed from f32 to f64
- Split Float/Real pattern matches in formatting.rs since they now have
  different underlying types (f32 vs f64)
- Merge Real with Double in sqllogictest test files (both are f64)
- Downgrade reqwest from 0.13 to 0.12 in vibesql-server to match the
  version used by opendal and avoid API incompatibilities

+++ crates/vibesql-executor/tests/conversion_function_tests.rs
@@ -90,7 +90,7 @@ fn float_lit(value: f32) -> vibesql_ast::Expression {
 }
 
 /// Helper to create a real literal expression
-fn real_lit(value: f32) -> vibesql_ast::Expression {
+fn real_lit(value: f64) -> vibesql_ast::Expression {
     vibesql_ast::Expression::Literal(vibesql_types::SqlValue::Real(value))
 }
